Split pages & components, updated layout design, minor error fixes
