+++ src/components/Account/UpdatePassword/index.js
@@ -4,7 +4,6 @@ import { compose } from 'recompose';
 import Alert from '@material-ui/lab/Alert';
 import Box from '@material-ui/core/Box';
 import Button from '@material-ui/core/Button';
-import Container from '@material-ui/core/Container';
 import Paper from '@material-ui/core/Paper';
 import Snackbar from '@material-ui/core/Snackbar';
 import TextField from '@material-ui/core/TextField';
@@ -12,7 +11,7 @@ import Typography from '@material-ui/core/Typography';
 
 import { withStyles } from '@material-ui/core/styles';
  
-import { withFirebase } from '../../firebase';
+import { withFirebase } from '../../../firebase';
 
 const styles = theme => ({
   form: {
@@ -24,12 +23,13 @@ const styles = theme => ({
 });
 
 const INITIAL_STATE = {
-  email: '',
+  passwordOne: '',
+  passwordTwo: '',
   success: null,
   error: null,
 };
 
-class UpdateEmailBase extends Component {
+class UpdatePasswordBase extends Component {
   constructor(props) {
     super(props);
  
@@ -45,15 +45,12 @@ class UpdateEmailBase extends Component {
   }
 
   onSubmit = event => {
-    const { email } = this.state;
+    const { passwordOne } = this.state;
  
     this.props.firebase
-      .doUpdateEmail(email)
+      .doUpdatePassword(passwordOne)
       .then(() => {
-        return this.props.firebase.doSendEmailVerification();
-      })
-      .then(() => {
-        let success = { code: 200, message: "Your email has been updated. Check inbox for verification email." };
+        let success = { code: 200, message: "Your password has been updated." };
         this.setState({ success });
       })
       .catch(error => {
@@ -74,9 +71,10 @@ class UpdateEmailBase extends Component {
   render() {
     const { classes } = this.props;
 
-    const { email, success, error } = this.state;
- 
-    const isInvalid = email === '';
+    const { passwordOne, passwordTwo, success, error } = this.state;
+
+    const isInvalid = passwordOne !== passwordTwo ||
+                      passwordOne === '';
 
     const isSuccess = success !== null;
 
@@ -86,46 +84,53 @@ class UpdateEmailBase extends Component {
  
     return (
       <React.Fragment>
-        <Container maxWidth="sm" disableGutters>
-          <Box py={3}>
-            <Paper elevation={0}>
-              <Box px={3} pt={3}>
-                <Typography align="center" variant="h4">    
-                  <strong>Manage Email</strong>
-                </Typography>
-              </Box>
-
-              <Box p={3}>
-                <form className={classes.form} onSubmit={this.onSubmit}>
-                  <TextField
-                    error={isError}
-                    fullWidth
-                    id="email"
-                    helperText="You'll need to confirm that this email belongs to you."
-                    label="Email Address"
-                    margin="normal"
-                    name="email"
-                    onChange={this.onChange}
-                    required
-                    value={email}
-                    variant="filled"
-                  />
-                  <Button
-                    className={classes.submit}
-                    color="primary"
-                    disabled={isDisabled}
-                    fullWidth
-                    size="large"
-                    type="submit"
-                    variant="contained"
-                  >
-                    Update My Email
-                  </Button>
-                </form>
-              </Box>
-            </Paper>
+        <Paper elevation={0}>
+          <Box p={3}>
+            <Typography align="center" variant="h4" gutterBottom>    
+              <strong>Manage Password</strong>
+            </Typography>
+            
+            <form className={classes.form} onSubmit={this.onSubmit}>
+              <TextField
+                error={isError}
+                fullWidth
+                id="passwordOne"
+                label="Password"
+                margin="normal"
+                name="passwordOne"
+                onChange={this.onChange}
+                required
+                type="password"
+                value={passwordOne}
+                variant="filled"
+              />
+              <TextField
+                error={isError}
+                fullWidth
+                id="passwordTwo"
+                label="Confirm Password"
+                margin="normal"
+                name="passwordTwo"
+                onChange={this.onChange}
+                required
+                type="password"
+                value={passwordTwo}
+                variant="filled"
+              />
+              <Button
+                className={classes.submit}
+                color="primary"
+                disabled={isDisabled}
+                fullWidth
+                size="large"
+                type="submit"
+                variant="contained"
+              >
+                Update My Password
+              </Button>
+            </form>
           </Box>
-        </Container>
+        </Paper>
 
         {success &&
           <Snackbar open={isSuccess} autoHideDuration={6000} onClose={this.handleClose}>
@@ -147,9 +152,9 @@ class UpdateEmailBase extends Component {
   }
 }
 
-const UpdateEmail = compose(
+const UpdatePassword = compose(
   withStyles(styles, { withTheme: true }),
   withFirebase,
-)(UpdateEmailBase);
+)(UpdatePasswordBase);
  
-export default UpdateEmail;
+export default withFirebase(UpdatePassword);
