+++ src/components/Account/index.js
@@ -1,0 +1,5 @@
+import UpdateEmail from './UpdateEmail';
+import UpdatePassword from './UpdatePassword';
+import UpdateProfile from './UpdateProfile';
+
+export { UpdateEmail, UpdatePassword, UpdateProfile };
