+++ src/components/Action/index.js
@@ -1,0 +1,5 @@
+import RecoverEmail from './RecoverEmail';
+import ResetPassword from './ResetPassword';
+import VerifyEmail from './VerifyEmail';
+
+export { RecoverEmail, ResetPassword, VerifyEmail };
