+++ src/components/PasswordForget/index.js
@@ -1,0 +1,4 @@
+import PasswordForgetForm from './passwordForgetForm';
+import PasswordForgetLink from './passwordForgetLink';
+
+export { PasswordForgetForm, PasswordForgetLink };
