+++ src/components/SignIn/index.js
@@ -1,0 +1,4 @@
+import SignInForm from './signInForm';
+import SignInLink from './signInLink';
+
+export { SignInForm, SignInLink };
