+++ src/components/SignUp/index.js
@@ -1,0 +1,4 @@
+import SignUpForm from './signUpForm';
+import SignUpLink from './signUpLink';
+
+export { SignUpForm, SignUpLink };
