+++ src/pages/Home/index.js
@@ -12,7 +12,7 @@ class HomeBase extends Component {
   render() {
     return(
       <Container maxWidth="sm">
-        <Box py={3}>
+        <Box pt={2}>
           <Paper elevation={0}>
             <Box p={3}>
               <Typography align="center" variant="h4" gutterBottom>
